# TextServerFallback: 2x performance improvement by removing redundant lookups