followup to https://github.com/godotengine/godot/pull/92575

- removes redundant FontForSizeFallback cache lookups
- also caches editor setting
- usies a faster hash method on the FontForSizeFallback cache (its the same function its just not converting the key to a variant before hashing)
- ShapedTextDataFallback::valid changed to an atomic bool to avoid locking a mutex, as its read Very often

in total from my benchmark it takes ~38% of the time to render the same text in a Label (~44% in CodeEdit, the main script editor in godot)

benchmark: [TextServerTest.zip](https://github.com/user-attachments/files/15508175/TextServerTest.zip)
benchmark compile flags: `scons target=template_release production=yes optimize=speed lto=full tests=yes deprecated=no scu_build=yes module_text_server_adv_enabled=no module_text_server_fb_enabled=yes`
benchmark run command: `./godot/bin/godot.linuxbsd.template_release.x86_64 --disable-vsync ./TextServerTest/project.godot`

results before:
```
Godot Engine v4.3.beta.custom_build.6b38be699 (2024-05-30 20:38:39 UTC) - https://godotengine.org
Vulkan 1.3.277 - Forward+ - Using Device #0: NVIDIA - NVIDIA GeForce GTX 1080 Ti

18.2138115468651ms avg draw time (99 samples)
18.2687489673345ms avg draw time (198 samples)
18.1693966541226ms avg draw time (297 samples)
18.088425048674ms avg draw time (396 samples)
18.0387193506414ms avg draw time (495 samples)
18.0871025079027ms avg draw time (594 samples)
18.1229936784136ms avg draw time (693 samples)
18.1424099989612ms avg draw time (792 samples)
18.1100280196578ms avg draw time (891 samples)
18.088856369558ms avg draw time (990 samples)
18.1047671206836ms avg draw time (1089 samples)
```
after:
```
Godot Engine v4.3.beta.custom_build.7a4461225 (2024-05-31 05:57:28 UTC) - https://godotengine.org
Vulkan 1.3.277 - Forward+ - Using Device #0: NVIDIA - NVIDIA GeForce GTX 1080 Ti

6.8193662046182ms avg draw time (99 samples)
6.81492054101193ms avg draw time (198 samples)
6.8215898391775ms avg draw time (297 samples)
6.8254777879426ms avg draw time (396 samples)
6.82788376856332ms avg draw time (495 samples)
6.80477530867965ms avg draw time (594 samples)
6.78739540848725ms avg draw time (693 samples)
6.77584548189182ms avg draw time (792 samples)
6.76721113699454ms avg draw time (891 samples)
6.75983934691458ms avg draw time (990 samples)
6.75281939756115ms avg draw time (1089 samples)
```


the last line of each is the most important, its the average of all frame times (minus some longer startup frames)